requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-cors
flask-restful
datetime 
celery_async.celery_async_functions 
flask_jwt_extended

